requirements.txt
numpy==1.16.4
Pillow==7.1.1
tqdm==4.45.0
